refactor(citation): clean up state management of citation type
