+++ addon/components/citation-plugin/citation-insert.hbs
@@ -10,12 +10,13 @@
   </AuButton>
 </AuList::Item>
 
-{{#if this.showModal}}
 <CitationPlugin::Citations::SearchModal
+  @open={{this.showModal}}
   @closeModal={{this.closeModal}}
   @insertDecisionCitation={{this.insertDecisionCitation}}
   @insertArticleCitation={{this.insertArticleCitation}}
-  @legislationTypeUri={{this.legislationTypeUri}}
+  @legislationTypeUri={{this.selectedLegislationTypeUri}}
+
+  @selectLegislationType={{this.selectLegislationType}}
   @text={{this.text}}
 />
-{{/if}}
